+++ core/src/main/java/org/stellar/anchor/ledger/LedgerClientHelper.java
@@ -2,27 +2,18 @@ package org.stellar.anchor.ledger;
 
 import static java.lang.Thread.sleep;
 import static org.stellar.anchor.ledger.LedgerTransaction.*;
-import static org.stellar.anchor.util.AssetHelper.toXdrAmount;
 import static org.stellar.anchor.util.Log.*;
 import static org.stellar.sdk.responses.sorobanrpc.SendTransactionResponse.*;
 import static org.stellar.sdk.xdr.HostFunctionType.HOST_FUNCTION_TYPE_INVOKE_CONTRACT;
 import static org.stellar.sdk.xdr.OperationType.*;
 import static org.stellar.sdk.xdr.SignerKeyType.*;
-import static org.stellar.sdk.xdr.SignerKeyType.SIGNER_KEY_TYPE_ED25519_SIGNED_PAYLOAD;
 
-import java.io.IOException;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
 import org.stellar.anchor.api.exception.LedgerException;
 import org.stellar.sdk.StrKey;
 import org.stellar.sdk.TOID;
-import org.stellar.sdk.responses.SubmitTransactionAsyncResponse;
-import org.stellar.sdk.responses.operations.OperationResponse;
-import org.stellar.sdk.responses.operations.PathPaymentBaseOperationResponse;
-import org.stellar.sdk.responses.operations.PaymentOperationResponse;
-import org.stellar.sdk.responses.sorobanrpc.GetTransactionResponse;
-import org.stellar.sdk.responses.sorobanrpc.GetTransactionsResponse;
 import org.stellar.sdk.xdr.*;
 
 public class LedgerClientHelper {
@@ -194,32 +185,6 @@ public class LedgerClientHelper {
 
   public record ParseResult(Operation[] operations, String sourceAccount, Memo memo) {}
 
-  /**
-   * Convert from Horizon signer key type to XDR signer key type.
-   *
-   * @param type the Horizon signer key type
-   * @return the XDR signer key type
-   */
-  public static SignerKeyType getKeyTypeDiscriminant(String type) {
-    return switch (type) {
-      case "ed25519_public_key" -> SIGNER_KEY_TYPE_ED25519;
-      case "preauth_tx" -> SIGNER_KEY_TYPE_PRE_AUTH_TX;
-      case "sha256_hash" -> SIGNER_KEY_TYPE_HASH_X;
-      case "ed25519_signed_payload" -> SIGNER_KEY_TYPE_ED25519_SIGNED_PAYLOAD;
-      default -> throw new IllegalArgumentException("Invalid signer key type: " + type);
-    };
-  }
-
-  public static SendTransactionStatus convert(
-      SubmitTransactionAsyncResponse.TransactionStatus status) {
-    return switch (status) {
-      case PENDING -> SendTransactionStatus.PENDING;
-      case ERROR -> SendTransactionStatus.ERROR;
-      case DUPLICATE -> SendTransactionStatus.DUPLICATE;
-      case TRY_AGAIN_LATER -> SendTransactionStatus.TRY_AGAIN_LATER;
-    };
-  }
-
   public static List<LedgerOperation> getLedgerOperations(
       Integer applicationOrder, Long sequenceNumber, ParseResult parseResult)
       throws LedgerException {
@@ -269,108 +234,5 @@ public class LedgerClientHelper {
       info("Interrupted while waiting for transaction to complete");
     }
     throw new LedgerException("Transaction took too long to complete");
-  }
-
-  public static LedgerOperation toLedgerOperation(OperationResponse op) {
-    LedgerOperation.LedgerOperationBuilder builder = LedgerOperation.builder();
-    // TODO: Capture muxed account events
-    if (op instanceof PaymentOperationResponse paymentOp) {
-      builder.type(PAYMENT);
-      builder.paymentOperation(
-          LedgerTransaction.LedgerPaymentOperation.builder()
-              .id(String.valueOf(paymentOp.getId()))
-              .from(paymentOp.getFrom())
-              .to(paymentOp.getTo())
-              .amount(toXdrAmount(paymentOp.getAmount()))
-              .asset(paymentOp.getAsset().toXdr())
-              .sourceAccount(paymentOp.getSourceAccount())
-              .build());
-    } else if (op instanceof PathPaymentBaseOperationResponse pathPaymentOp) {
-      builder.type(PATH_PAYMENT_STRICT_RECEIVE);
-      builder.pathPaymentOperation(
-          LedgerTransaction.LedgerPathPaymentOperation.builder()
-              .id(String.valueOf(pathPaymentOp.getId()))
-              .from(pathPaymentOp.getFrom())
-              .to(pathPaymentOp.getTo())
-              .amount(toXdrAmount(pathPaymentOp.getAmount()))
-              .asset(pathPaymentOp.getAsset().toXdr())
-              .sourceAccount(pathPaymentOp.getSourceAccount())
-              .build());
-    } else {
-      return null;
-    }
-    return builder.build();
-  }
-
-  /**
-   * Convert a GetTransactionResponse to a LedgerTransaction.
-   *
-   * @param txnResponse the GetTransactionResponse to convert
-   * @return the converted LedgerTransaction
-   * @throws LedgerException if the transaction is null or malformed
-   */
-  public static LedgerTransaction fromGetTransactionResponse(GetTransactionResponse txnResponse)
-      throws LedgerException {
-    TransactionEnvelope txnEnv;
-    try {
-      txnEnv = TransactionEnvelope.fromXdrBase64(txnResponse.getEnvelopeXdr());
-    } catch (IOException ioex) {
-      throw new LedgerException("Unable to parse transaction envelope", ioex);
-    }
-    Integer applicationOrder = txnResponse.getApplicationOrder();
-    Long sequenceNumber = txnResponse.getLedger();
-    ParseResult parseResult =
-        LedgerClientHelper.parseOperationAndSourceAccountAndMemo(txnEnv, txnResponse.getTxHash());
-    if (parseResult == null) return null;
-    List<LedgerTransaction.LedgerOperation> operations =
-        LedgerClientHelper.getLedgerOperations(applicationOrder, sequenceNumber, parseResult);
-
-    return LedgerTransaction.builder()
-        .hash(txnResponse.getTxHash())
-        .ledger(txnResponse.getLedger())
-        .applicationOrder(txnResponse.getApplicationOrder())
-        .sourceAccount(parseResult.sourceAccount())
-        .envelopeXdr(txnResponse.getEnvelopeXdr())
-        .memo(parseResult.memo())
-        .sequenceNumber(sequenceNumber)
-        .createdAt(Instant.ofEpochSecond(txnResponse.getCreatedAt()))
-        .operations(operations)
-        .build();
-  }
-
-  /**
-   * Converting from Stellar RPC transaction to LedgerTransaction. TODO: This function will be
-   * removed after migrating to getEvents methods.
-   *
-   * @param txn the Stellar RPC transaction to convert
-   * @return the converted LedgerTransaction
-   * @throws LedgerException if the transaction is null or malformed
-   */
-  public static LedgerTransaction fromStellarRpcTransaction(GetTransactionsResponse.Transaction txn)
-      throws LedgerException {
-    TransactionEnvelope txnEnv;
-    try {
-      txnEnv = TransactionEnvelope.fromXdrBase64(txn.getEnvelopeXdr());
-    } catch (IOException ioex) {
-      throw new LedgerException("Unable to parse transaction envelope", ioex);
-    }
-    Integer applicationOrder = txn.getApplicationOrder();
-    Long sequenceNumber = txn.getLedger();
-    ParseResult parseResult =
-        LedgerClientHelper.parseOperationAndSourceAccountAndMemo(txnEnv, txn.getTxHash());
-    if (parseResult == null) return null;
-    List<LedgerTransaction.LedgerOperation> operations =
-        LedgerClientHelper.getLedgerOperations(applicationOrder, sequenceNumber, parseResult);
-    return LedgerTransaction.builder()
-        .hash(txn.getTxHash())
-        .ledger(txn.getLedger())
-        .applicationOrder(txn.getApplicationOrder())
-        .sourceAccount(parseResult.sourceAccount())
-        .envelopeXdr(txn.getEnvelopeXdr())
-        .memo(parseResult.memo())
-        .sequenceNumber(sequenceNumber)
-        .createdAt(Instant.ofEpochSecond(txn.getCreatedAt()))
-        .operations(operations)
-        .build();
   }
 }
