+++ core/src/main/java/org/stellar/anchor/ledger/StellarRpc.java
@@ -1,10 +1,10 @@
 package org.stellar.anchor.ledger;
 
-import static org.stellar.anchor.ledger.LedgerClientHelper.fromGetTransactionResponse;
 import static org.stellar.sdk.xdr.LedgerEntry.*;
 import static org.stellar.sdk.xdr.SignerKeyType.SIGNER_KEY_TYPE_ED25519;
 
 import java.io.IOException;
+import java.time.Instant;
 import java.util.*;
 import lombok.Getter;
 import lombok.SneakyThrows;
@@ -111,20 +111,6 @@ public class StellarRpc implements LedgerClient {
         .build();
   }
 
-  @SneakyThrows
-  public static AccountEntry getAccountEntry(SorobanServer sorobanServer, String accountId) {
-    KeyPair kp = KeyPair.fromAccountId(accountId);
-    List<LedgerKey> ledgerKeys =
-        Collections.singletonList(
-            LedgerKey.builder()
-                .account(LedgerKeyAccount.builder().accountID(kp.getXdrAccountId()).build())
-                .discriminant(LedgerEntryType.ACCOUNT)
-                .build());
-    GetLedgerEntriesResponse response = sorobanServer.getLedgerEntries(ledgerKeys);
-    return LedgerEntry.LedgerEntryData.fromXdrBase64(response.getEntries().get(0).getXdr())
-        .getAccount();
-  }
-
   private TrustLineEntry getTrustlineRpc(String accountId, String asset) throws LedgerException {
     KeyPair kp = KeyPair.fromAccountId(accountId);
 
@@ -166,5 +152,41 @@ public class StellarRpc implements LedgerClient {
     LedgerEntryData ledgerEntryData =
         LedgerEntryData.fromXdrBase64(response.getEntries().get(0).getXdr());
     return ledgerEntryData.getAccount();
+  }
+
+  /**
+   * Convert a GetTransactionResponse to a LedgerTransaction.
+   *
+   * @param txnResponse the GetTransactionResponse to convert
+   * @return the converted LedgerTransaction
+   * @throws LedgerException if the transaction is null or malformed
+   */
+  public static LedgerTransaction fromGetTransactionResponse(GetTransactionResponse txnResponse)
+      throws LedgerException {
+    TransactionEnvelope txnEnv;
+    try {
+      txnEnv = TransactionEnvelope.fromXdrBase64(txnResponse.getEnvelopeXdr());
+    } catch (IOException ioex) {
+      throw new LedgerException("Unable to parse transaction envelope", ioex);
+    }
+    Integer applicationOrder = txnResponse.getApplicationOrder();
+    Long sequenceNumber = txnResponse.getLedger();
+    LedgerClientHelper.ParseResult parseResult =
+        LedgerClientHelper.parseOperationAndSourceAccountAndMemo(txnEnv, txnResponse.getTxHash());
+    if (parseResult == null) return null;
+    List<LedgerTransaction.LedgerOperation> operations =
+        LedgerClientHelper.getLedgerOperations(applicationOrder, sequenceNumber, parseResult);
+
+    return LedgerTransaction.builder()
+        .hash(txnResponse.getTxHash())
+        .ledger(txnResponse.getLedger())
+        .applicationOrder(txnResponse.getApplicationOrder())
+        .sourceAccount(parseResult.sourceAccount())
+        .envelopeXdr(txnResponse.getEnvelopeXdr())
+        .memo(parseResult.memo())
+        .sequenceNumber(sequenceNumber)
+        .createdAt(Instant.ofEpochSecond(txnResponse.getCreatedAt()))
+        .operations(operations)
+        .build();
   }
 }
