+++ core/src/test/kotlin/org/stellar/anchor/ledger/LedgerClientHelperTest.kt
@@ -3,9 +3,7 @@ package org.stellar.anchor.ledger
 import kotlin.test.assertEquals
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
 import org.stellar.anchor.util.GsonUtils
-import org.stellar.sdk.responses.SubmitTransactionAsyncResponse.TransactionStatus
 import org.stellar.sdk.responses.sorobanrpc.SendTransactionResponse.SendTransactionStatus.*
 import org.stellar.sdk.xdr.*
 import org.stellar.sdk.xdr.CryptoKeyType.KEY_TYPE_ED25519
@@ -15,40 +13,6 @@ import org.stellar.sdk.xdr.OperationType.PATH_PAYMENT_STRICT_RECEIVE
 import org.stellar.sdk.xdr.SignerKeyType.*
 
 internal class LedgerClientHelperTest {
-  @Test
-  fun `test getKeyTypeDiscriminant with valid types`() {
-    assertEquals(
-      SIGNER_KEY_TYPE_ED25519,
-      LedgerClientHelper.getKeyTypeDiscriminant("ed25519_public_key"),
-    )
-    assertEquals(
-      SIGNER_KEY_TYPE_PRE_AUTH_TX,
-      LedgerClientHelper.getKeyTypeDiscriminant("preauth_tx"),
-    )
-    assertEquals(SIGNER_KEY_TYPE_HASH_X, LedgerClientHelper.getKeyTypeDiscriminant("sha256_hash"))
-    assertEquals(
-      SIGNER_KEY_TYPE_ED25519_SIGNED_PAYLOAD,
-      LedgerClientHelper.getKeyTypeDiscriminant("ed25519_signed_payload"),
-    )
-  }
-
-  @Test
-  fun `test getKeyTypeDiscriminant with invalid type`() {
-    val exception =
-      assertThrows<IllegalArgumentException> {
-        LedgerClientHelper.getKeyTypeDiscriminant("invalid_type")
-      }
-    assertEquals("Invalid signer key type: invalid_type", exception.message)
-  }
-
-  @Test
-  fun `test convert() with transaction status`() {
-    assertEquals(PENDING, LedgerClientHelper.convert(TransactionStatus.PENDING))
-    assertEquals(ERROR, LedgerClientHelper.convert(TransactionStatus.ERROR))
-    assertEquals(DUPLICATE, LedgerClientHelper.convert(TransactionStatus.DUPLICATE))
-    assertEquals(TRY_AGAIN_LATER, LedgerClientHelper.convert(TransactionStatus.TRY_AGAIN_LATER))
-  }
-
   @Test
   fun `test convert() with payment transaction`() {
     val operation = GsonUtils.getInstance().fromJson(testPaymentOpJson, Operation::class.java)
@@ -242,7 +206,7 @@ internal class LedgerClientHelperTest {
   }
 }
 
-private val testPaymentOpJson =
+private const val testPaymentOpJson =
   """
 {
    "body":{
@@ -265,7 +229,7 @@ private val testPaymentOpJson =
 }
 """
 
-private val testPathPaymentOpJson =
+private const val testPathPaymentOpJson =
   """
 {
    "body":{
@@ -294,7 +258,8 @@ private val testPathPaymentOpJson =
 }
 """
 
-private val testUnhandledOpJson = """
+private const val testUnhandledOpJson =
+  """
 {
    "body":{
       "discriminant":"CREATE_ACCOUNT"
