+++ platform/src/main/java/org/stellar/anchor/platform/observer/stellar/SacToAssetMapper.java
@@ -3,5 +3,7 @@ package org.stellar.anchor.platform.observer.stellar;
 import org.stellar.sdk.xdr.Asset;
 
 public class SacToAssetMapper {
-  public static Asset map(String sac) {}
+  public static Asset map(String sac) {
+    return null;
+  }
 }
